Moved pixel grouping checking out of DFS and put into its own GroupingService
