+++ tests/unit/backend/data/test_DataFactoryService.py
@@ -430,187 +430,3 @@ class TestDataFactoryService(unittest.TestCase):
     def test_getLiveMetadata(self):
         actual = self.instance.getLiveMetadata()
         assert actual == self.expected()
-
-    ##### TEST FILTER CALIBRATION BY NON-STRADDLING GROUPS ####
-
-    def test_filterCalibrationRecordsByNonStraddlingGroups_keeps_valid(self):
-        from snapred.backend.dao.calibration.CalibrationRecord import CalibrationRecord
-        from snapred.backend.dao.calibration.FocusGroupMetric import FocusGroupMetric
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # Simulate no straddling groups
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = []
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        record = mock.MagicMock(spec=CalibrationRecord)
-        record.focusGroupCalibrationMetrics = mock.MagicMock(spec=FocusGroupMetric)
-        record.focusGroupCalibrationMetrics.focusGroupName = "Column"
-
-        result = self.instance.filterCalibrationRecordsByNonStraddlingGroups([record], "12345", True)
-        assert result == [record]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.assert_called_once_with(
-            "Checking panel straddle", GroupingWorkspace="fake_ws"
-        )
-
-    def test_filterCalibrationRecordsByNonStraddlingGroups_excludes_straddling(self):
-        from snapred.backend.dao.calibration.CalibrationRecord import CalibrationRecord
-        from snapred.backend.dao.calibration.FocusGroupMetric import FocusGroupMetric
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # Simulate straddling groups present
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = ["1", "2"]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        record = mock.MagicMock(spec=CalibrationRecord)
-        record.focusGroupCalibrationMetrics = mock.MagicMock(spec=FocusGroupMetric)
-        record.focusGroupCalibrationMetrics.focusGroupName = "All"
-        record.runNumber = "12345"
-        record.version = 1
-
-        result = self.instance.filterCalibrationRecordsByNonStraddlingGroups([record], "12345", True)
-        assert result == []
-
-    def test_filterCalibrationRecordsByNonStraddlingGroups_mixed(self):
-        from snapred.backend.dao.calibration.CalibrationRecord import CalibrationRecord
-        from snapred.backend.dao.calibration.FocusGroupMetric import FocusGroupMetric
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # First call (Column) -> no straddle, second call (All) -> straddles
-        callback1 = mock.MagicMock()
-        callback1.get.return_value = []
-        callback2 = mock.MagicMock()
-        callback2.get.return_value = ["1"]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.side_effect = [callback1, callback2]
-
-        record_good = mock.MagicMock(spec=CalibrationRecord)
-        record_good.focusGroupCalibrationMetrics = mock.MagicMock(spec=FocusGroupMetric)
-        record_good.focusGroupCalibrationMetrics.focusGroupName = "Column"
-
-        record_bad = mock.MagicMock(spec=CalibrationRecord)
-        record_bad.focusGroupCalibrationMetrics = mock.MagicMock(spec=FocusGroupMetric)
-        record_bad.focusGroupCalibrationMetrics.focusGroupName = "All"
-        record_bad.runNumber = "12345"
-        record_bad.version = 2
-
-        result = self.instance.filterCalibrationRecordsByNonStraddlingGroups([record_good, record_bad], "12345", True)
-        assert result == [record_good]
-
-    def test_filterCalibrationRecordsByNonStraddlingGroups_caches_per_groupname(self):
-        from snapred.backend.dao.calibration.CalibrationRecord import CalibrationRecord
-        from snapred.backend.dao.calibration.FocusGroupMetric import FocusGroupMetric
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = []
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        records = []
-        for _ in range(3):
-            r = mock.MagicMock(spec=CalibrationRecord)
-            r.focusGroupCalibrationMetrics = mock.MagicMock(spec=FocusGroupMetric)
-            r.focusGroupCalibrationMetrics.focusGroupName = "Column"
-            records.append(r)
-
-        result = self.instance.filterCalibrationRecordsByNonStraddlingGroups(records, "12345", True)
-        assert len(result) == 3
-        # Only one call to the algorithm despite three records with the same grouping
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.assert_called_once()
-
-    def test_filterNormalizationRecordsByNonStraddlingGroups_keeps_valid(self):
-        from snapred.backend.dao.normalization.NormalizationRecord import NormalizationRecord
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # Simulate no straddling groups
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = []
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        record = mock.MagicMock(spec=NormalizationRecord)
-        record.calculationParameters = mock.MagicMock()
-        record.calculationParameters.name = "Column"
-
-        result = self.instance.filterNormalizationRecordsByNonStraddlingGroups([record], "12345", True)
-        assert result == [record]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.assert_called_once_with(
-            "Checking panel straddle", GroupingWorkspace="fake_ws"
-        )
-
-    def test_filterNormalizationRecordsByNonStraddlingGroups_excludes_straddling(self):
-        from snapred.backend.dao.normalization.NormalizationRecord import NormalizationRecord
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # Simulate straddling groups present
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = ["1", "2"]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        record = mock.MagicMock(spec=NormalizationRecord)
-        record.calculationParameters = mock.MagicMock()
-        record.calculationParameters.name = "All"
-        record.runNumber = "12345"
-        record.version = 1
-
-        result = self.instance.filterNormalizationRecordsByNonStraddlingGroups([record], "12345", True)
-        assert result == []
-
-    def test_filterNormalizationRecordsByNonStraddlingGroups_mixed(self):
-        from snapred.backend.dao.normalization.NormalizationRecord import NormalizationRecord
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        # First call (Column) -> no straddle, second call (All) -> straddles
-        callback1 = mock.MagicMock()
-        callback1.get.return_value = []
-        callback2 = mock.MagicMock()
-        callback2.get.return_value = ["1"]
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.side_effect = [callback1, callback2]
-
-        record_good = mock.MagicMock(spec=NormalizationRecord)
-        record_good.calculationParameters = mock.MagicMock()
-        record_good.calculationParameters.name = "Column"
-
-        record_bad = mock.MagicMock(spec=NormalizationRecord)
-        record_bad.calculationParameters = mock.MagicMock()
-        record_bad.calculationParameters.name = "All"
-        record_bad.runNumber = "12345"
-        record_bad.version = 2
-
-        result = self.instance.filterNormalizationRecordsByNonStraddlingGroups([record_good, record_bad], "12345", True)
-        assert result == [record_good]
-
-    def test_filterNormalizationRecordsByNonStraddlingGroups_caches_per_groupname(self):
-        from snapred.backend.dao.normalization.NormalizationRecord import NormalizationRecord
-
-        self.instance.mantidSnapper = mock.MagicMock()
-        self.instance.groceryService = mock.MagicMock()
-        self.instance.groceryService.fetchGroupingDefinition.return_value = {"workspace": "fake_ws"}
-        straddlingCallback = mock.MagicMock()
-        straddlingCallback.get.return_value = []
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.return_value = straddlingCallback
-
-        records = []
-        for _ in range(3):
-            r = mock.MagicMock(spec=NormalizationRecord)
-            r.calculationParameters = mock.MagicMock()
-            r.calculationParameters.name = "Column"
-            records.append(r)
-
-        result = self.instance.filterNormalizationRecordsByNonStraddlingGroups(records, "12345", True)
-        assert len(result) == 3
-        # Only one call to the algorithm despite three records with the same grouping
-        self.instance.mantidSnapper.DetectorPanelStraddleCheck.assert_called_once()
